Changed output of get_duplicates()
Changed return value of get_duplicates() of a list of dictionary-type values representing the duplicate entries to a nested dictionary, with the key being the english word and the value being a list of dictionary-type values for the duplicate types.

That way, it is easier to iterate through them and it is a more sensible data structure to store them under.