Data structures changed
Changing get_dictionary() output from a list to a defaultdict to line up with RtfParser a bit more.

This helps performance too and cleans up the code. Lookup (x in dict) is O(1) instead of O(N) with a dict rather than a list.

I also redid get_briefs and _missing_words to handle the new output.